Split profile sections into individual components and actions
